✨ add traffic::borrowed_mutex(TSCont) constructor
While the continuation type we have allows for it to be used "as a
lock", we haven't correctly wrapped the type in its entirety (and it
will be some time before we do, as we need to add wrapping callbacks
correctly first). However in the interim, we can build a borrowed_mutex
from a TSCont.